requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
bcrypt==4.0.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
faiss-cpu==1.13.0
fastapi==0.135.2
filelock==3.25.2
fsspec==2026.3.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
openai==2.30.0
packaging==26.0
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.3.32
requests==2.33.0
rich==14.3.3
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==81.0.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0